alteracao do nome de sofascore para provider
