+++ provider/src/main/java/com/sevenine/provider/SofascoreApplication.java
@@ -1,0 +1,15 @@
+package com.sevenine.provider;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@EnableMongoRepositories
+@SpringBootApplication
+public class SofascoreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SofascoreApplication.class, args);
+    }
+
+}
