+++ provider/src/main/java/com/sevenine/provider/application/service/TeamApiService.java
@@ -1,0 +1,9 @@
+package com.sevenine.provider.application.service;
+
+import com.sevenine.provider.application.service.response.TeamResponse;
+
+import java.util.List;
+
+public interface TeamApiService {
+    List<TeamResponse> list(String tournamentCode, String seasonCode);
+}
