+++ provider/src/main/java/com/sevenine/provider/application/service/response/TeamResponse.java
@@ -1,0 +1,4 @@
+package com.sevenine.provider.application.service.response;
+
+public record TeamResponse(int id, String name, String slug, String shortName, String nameCode) {
+}
