+++ provider/src/main/java/com/sevenine/provider/application/usecase/UpdateTeamUsecase.java
@@ -1,0 +1,27 @@
+package com.sevenine.provider.application.usecase;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sevenine.provider.application.service.TeamApiService;
+import com.sevenine.provider.application.service.response.TeamResponse;
+import com.sevenine.provider.application.usecase.input.UpdateTeamInput;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Component
+public class UpdateTeamUsecase {
+
+    private final TeamApiService teamApiService;
+
+    private final ObjectMapper mapper;
+
+    public void execute(UpdateTeamInput input) {
+        List<TeamResponse> response =
+                teamApiService.list(input.tournament().getCode(), input.season().getCode());
+
+        System.out.println(response);
+    }
+
+}
