+++ provider/src/main/java/com/sevenine/provider/business/enumerated/ProviderEnum.java
@@ -1,0 +1,5 @@
+package com.sevenine.provider.business.enumerated;
+
+public enum ProviderEnum {
+    SOFASCORE
+}
