+++ provider/src/main/java/com/sevenine/provider/business/enumerated/SeasonEnum.java
@@ -1,0 +1,12 @@
+package com.sevenine.provider.business.enumerated;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum SeasonEnum {
+    YEAR2024("57170");
+
+    private final String code;
+}
