+++ provider/src/main/java/com/sevenine/provider/infrastructure/client/apidojo/response/RowsApiDojo.java
@@ -1,0 +1,7 @@
+package com.sevenine.provider.infrastructure.client.apidojo.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record RowsApiDojo(TeamApiDojo team) {
+}
