+++ provider/src/main/java/com/sevenine/provider/infrastructure/client/apidojo/response/StandingsApiDojo.java
@@ -1,0 +1,9 @@
+package com.sevenine.provider.infrastructure.client.apidojo.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record StandingsApiDojo(List<RowsApiDojo> rows) {
+}
