+++ provider/src/main/java/com/sevenine/provider/infrastructure/client/apidojo/response/TotalApiDojo.java
@@ -1,0 +1,9 @@
+package com.sevenine.provider.infrastructure.client.apidojo.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record TotalApiDojo(List<StandingsApiDojo> standings) {
+}
