+++ provider/src/main/java/com/sevenine/provider/infrastructure/client/apidojo/response/TournamentApiDojo.java
@@ -1,0 +1,10 @@
+package com.sevenine.provider.infrastructure.client.apidojo.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class TournamentApiDojo {
+    private String standings;
+}
