+++ settings.gradle
@@ -2,5 +2,5 @@ rootProject.name = 'futebola-project'
 
 include 'auth'
 include 'player'
-include 'sofascore'
+include 'provider'
