--- sofascore/src/main/java/com/sevenine/sofascore/SofascoreApplication.java
@@ -1,15 +1,0 @@
-package com.sevenine.sofascore;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@EnableMongoRepositories
-@SpringBootApplication
-public class SofascoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SofascoreApplication.class, args);
-    }
-
-}
