--- sofascore/src/main/java/com/sevenine/sofascore/application/config/RestTemplateConfig.java
@@ -1,16 +1,0 @@
-package com.sevenine.sofascore.application.config;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder builder) {
-        return builder.build();
-    }
-
-}
