--- sofascore/src/main/java/com/sevenine/sofascore/application/service/TeamApiService.java
@@ -1,9 +1,0 @@
-package com.sevenine.sofascore.application.service;
-
-import com.sevenine.sofascore.application.service.response.TeamResponse;
-
-import java.util.List;
-
-public interface TeamApiService {
-    List<TeamResponse> list(String tournamentCode, String seasonCode);
-}
