--- sofascore/src/main/java/com/sevenine/sofascore/application/service/response/TeamResponse.java
@@ -1,4 +1,0 @@
-package com.sevenine.sofascore.application.service.response;
-
-public record TeamResponse(int id, String name, String slug, String shortName, String nameCode) {
-}
