--- sofascore/src/main/java/com/sevenine/sofascore/business/enumerated/ProviderEnum.java
@@ -1,5 +1,0 @@
-package com.sevenine.sofascore.business.enumerated;
-
-public enum ProviderEnum {
-    SOFASCORE
-}
