--- sofascore/src/main/java/com/sevenine/sofascore/business/enumerated/SeasonEnum.java
@@ -1,12 +1,0 @@
-package com.sevenine.sofascore.business.enumerated;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum SeasonEnum {
-    YEAR2024("57170");
-
-    private final String code;
-}
