--- sofascore/src/main/java/com/sevenine/sofascore/business/enumerated/TournamentEnum.java
@@ -1,12 +1,0 @@
-package com.sevenine.sofascore.business.enumerated;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum TournamentEnum {
-    PAULISTA("372"), BRASILEIRO("325");
-
-    private final String code;
-}
