--- sofascore/src/main/java/com/sevenine/sofascore/infrastructure/client/apidojo/response/RowsApiDojo.java
@@ -1,7 +1,0 @@
-package com.sevenine.sofascore.infrastructure.client.apidojo.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record RowsApiDojo(TeamApiDojo team) {
-}
