--- sofascore/src/main/java/com/sevenine/sofascore/infrastructure/client/apidojo/response/TotalApiDojo.java
@@ -1,9 +1,0 @@
-package com.sevenine.sofascore.infrastructure.client.apidojo.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import java.util.List;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record TotalApiDojo(List<StandingsApiDojo> standings) {
-}
