--- sofascore/src/main/java/com/sevenine/sofascore/infrastructure/client/apidojo/response/TournamentApiDojo.java
@@ -1,10 +1,0 @@
-package com.sevenine.sofascore.infrastructure.client.apidojo.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class TournamentApiDojo {
-    private String standings;
-}
