--- sofascore/src/main/java/com/sevenine/sofascore/infrastructure/repository/mongodb/TeamRepository.java
@@ -1,9 +1,0 @@
-package com.sevenine.sofascore.infrastructure.repository.mongodb;
-
-import com.sevenine.sofascore.infrastructure.repository.document.TeamDocument;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TeamRepository extends MongoRepository<TeamDocument, String> {
-}
